refactor: dropping `ExtendedNote`
When checking how note filters are used I noticed that the ExtendedNote type is no longer used so I dropped it in this PR.
