+++ yarn-project/aztec.js/src/api/note.ts
@@ -1,1 +1,1 @@
-export { UniqueNote, ExtendedNote, Comparator, Note } from '@aztec/stdlib/note';
+export { UniqueNote, Comparator, Note } from '@aztec/stdlib/note';
