+++ yarn-project/stdlib/src/note/index.ts
@@ -1,5 +1,5 @@
 export * from './comparator.js';
-export * from './extended_note.js';
+export * from './unique_note.js';
 export * from './notes_filter.js';
 export * from './note_status.js';
 export * from './note.js';
